#56: scope data element parsers and loader
