+++ app/lib/dfe_data_tables/data_element_parsers.rb
@@ -1,6 +1,6 @@
 # frozen_string_literal: true
 
 module DfEDataTables
-  module Parsers
+  module DataElementParsers
   end
 end
