+++ app/lib/dfe_data_tables/data_element_parsers/cin.rb
@@ -3,7 +3,7 @@
 require_relative 'sheet'
 
 module DfEDataTables
-  module Parsers
+  module DataElementParsers
     class Cin < Sheet
       private
