+++ app/lib/dfe_data_tables/data_element_parsers/exclusions_up_to2005.rb
@@ -3,7 +3,7 @@
 require_relative 'sheet'
 
 module DfEDataTables
-  module Parsers
+  module DataElementParsers
     class ExclusionsUpTo2005 < Sheet
       private
