+++ app/lib/dfe_data_tables/data_element_parsers/eyfsp.rb
@@ -3,7 +3,7 @@
 require_relative 'sheet'
 
 module DfEDataTables
-  module Parsers
+  module DataElementParsers
     class Eyfsp < Sheet
       private
