+++ app/lib/dfe_data_tables/data_elements_loader.rb
@@ -6,29 +6,29 @@ module DfEDataTables
       data_tables_workbook = Roo::Spreadsheet.open(data_tables_path)
 
       @sheets_to_process = [
-        DfEDataTables::Parsers::ScPupil.new(data_tables_workbook),
-        DfEDataTables::Parsers::ScAddresses.new(data_tables_workbook),
-        DfEDataTables::Parsers::PruCensus.new(data_tables_workbook),
-        DfEDataTables::Parsers::EarlyYearsCensus.new(data_tables_workbook),
-        DfEDataTables::Parsers::AltProvision.new(data_tables_workbook),
-        DfEDataTables::Parsers::ApAddresses.new(data_tables_workbook),
-        DfEDataTables::Parsers::Eyfsp.new(data_tables_workbook),
-        DfEDataTables::Parsers::Phonics.new(data_tables_workbook),
-        DfEDataTables::Parsers::Ks1.new(data_tables_workbook),
-        DfEDataTables::Parsers::Ks2.new(data_tables_workbook),
-        DfEDataTables::Parsers::Year7.new(data_tables_workbook),
-        DfEDataTables::Parsers::Ks3.new(data_tables_workbook),
-        DfEDataTables::Parsers::Ks4.new(data_tables_workbook),
-        DfEDataTables::Parsers::Ks5.new(data_tables_workbook),
-        DfEDataTables::Parsers::Cin.new(data_tables_workbook),
-        DfEDataTables::Parsers::Cla.new(data_tables_workbook),
-        DfEDataTables::Parsers::Absence.new(data_tables_workbook),
-        DfEDataTables::Parsers::ExclusionsUpTo2005.new(data_tables_workbook),
-        DfEDataTables::Parsers::ExclusionsFrom2005.new(data_tables_workbook),
-        DfEDataTables::Parsers::Plams.new(data_tables_workbook),
-        DfEDataTables::Parsers::Nccis.new(data_tables_workbook),
-        DfEDataTables::Parsers::Isp.new(data_tables_workbook),
-        DfEDataTables::Parsers::Ypmad.new(data_tables_workbook)
+        DfEDataTables::DataElementParsers::ScPupil.new(data_tables_workbook),
+        DfEDataTables::DataElementParsers::ScAddresses.new(data_tables_workbook),
+        DfEDataTables::DataElementParsers::PruCensus.new(data_tables_workbook),
+        DfEDataTables::DataElementParsers::EarlyYearsCensus.new(data_tables_workbook),
+        DfEDataTables::DataElementParsers::AltProvision.new(data_tables_workbook),
+        DfEDataTables::DataElementParsers::ApAddresses.new(data_tables_workbook),
+        DfEDataTables::DataElementParsers::Eyfsp.new(data_tables_workbook),
+        DfEDataTables::DataElementParsers::Phonics.new(data_tables_workbook),
+        DfEDataTables::DataElementParsers::Ks1.new(data_tables_workbook),
+        DfEDataTables::DataElementParsers::Ks2.new(data_tables_workbook),
+        DfEDataTables::DataElementParsers::Year7.new(data_tables_workbook),
+        DfEDataTables::DataElementParsers::Ks3.new(data_tables_workbook),
+        DfEDataTables::DataElementParsers::Ks4.new(data_tables_workbook),
+        DfEDataTables::DataElementParsers::Ks5.new(data_tables_workbook),
+        DfEDataTables::DataElementParsers::Cin.new(data_tables_workbook),
+        DfEDataTables::DataElementParsers::Cla.new(data_tables_workbook),
+        DfEDataTables::DataElementParsers::Absence.new(data_tables_workbook),
+        DfEDataTables::DataElementParsers::ExclusionsUpTo2005.new(data_tables_workbook),
+        DfEDataTables::DataElementParsers::ExclusionsFrom2005.new(data_tables_workbook),
+        DfEDataTables::DataElementParsers::Plams.new(data_tables_workbook),
+        DfEDataTables::DataElementParsers::Nccis.new(data_tables_workbook),
+        DfEDataTables::DataElementParsers::Isp.new(data_tables_workbook),
+        DfEDataTables::DataElementParsers::Ypmad.new(data_tables_workbook)
       ]
 
       process
